% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_fragment_size.R
\name{bin_fragment_size}
\alias{bin_fragment_size}
\title{Get histogram of fragment lengths from a bam file}
\usage{
bin_fragment_size(bam, mutations = NULL, tag = "", bin_size = 2,
  custom_bins = NULL, mutated_only = F, normalized = F,
  isProperPair = NA, min_size = 1, max_size = 400,
  ignore_trimmed = T, different_strands = T, simple_cigar = F)
}
\arguments{
\item{bam}{path to bam file.}

\item{mutations}{An optional data frame with mutations. Must have the columns CHROM, POS, REF, ALT.}

\item{tag}{the RG tag if the bam has more than one samplee.}

\item{bin_size}{the width of the bin (breaks) of the histogram.}

\item{custom_bins}{A numeric vector for custom breaks to bin the histogram of fragment length. Over-rides bin_size.}

\item{mutated_only}{A logical, whether to return the counts for only mutated reads. The 'mutations' input should be given when TRUE.}

\item{normalized}{A logical, whether to normalize the counts to the total number of reads.}

\item{isProperPair}{a logical wheter to return only proper pairs (T), only improper pairs (F), or it does not matter (NA).}

\item{min_size}{Integer with the lowest fragment length.}

\item{max_size}{Integer with the highest fragment length.}

\item{ignore_trimmed}{logical, whether to remove reads that have been hard trimmed.}

\item{different_strands}{logical, whether to keep only reads whose mates map to different strand.}

\item{simple_cigar}{logical, whether to include only reads with simple cigar.}
}
\value{
A data frame with one column corresponding to the sample name from the bam. 
Each row has the count of fragment lengths within the bin normalized by the total number of reads.
}
\description{
The function first extracts fragment length from a bam file then computes the histogram over defined bins. If normalized is TRUE, 
the counts per bin will be normalized to the total read counts. Optionally,
it can computes the histogram of fragment lengths only for mutated reads (confirmed ctDNA molecules).
}
