% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{compare_simulated_observed}
\alias{compare_simulated_observed}
\title{Compares simulated and observed}
\usage{
compare_simulated_observed(simulated, observed, depths)
}
\description{
Compares simulated and observed
}
