% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fragment_size_wps.R
\name{compute_fragment_size_wps}
\alias{compute_fragment_size_wps}
\title{Analysis of fragment ends}
\usage{
compute_fragment_size_wps(bam, targets, tag = "", window_size = 120,
  step_size = 5, min_size = 120, max_size = 180, ...)
}
\arguments{
\item{bam}{the input bam file}

\item{targets}{The targets to restrict the windows within. Must have the columns chr, start and end.
In case of whole-genome, specify full chromosomes targets.}

\item{tag}{the RG tag if the bam has more than one samplee.}

\item{window_size}{The window (bin) size to use within the targets}

\item{step_size}{The step size to use in case of overlapping bins.}

\item{min_size}{Restrict fragments to this minimum size.}

\item{max_size}{Restrict fragments to this maximum size.}

\item{...}{Other parameters passed to get_fragment_size}
}
\value{
a data frame with the first three columns having the bins coordinates and 
other columns having the WPS (raw and adjusted by coverage) and number of fragment ends (raw and adjusted by coverage).
}
\description{
Caluclates the number of fragment ends and the Windowed Protection Score (WPS) in genomic tiles within targets
}
\details{
Check Snyder et al., Cell 2016.
}
