% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_background_panel.R
\name{create_background_panel}
\alias{create_background_panel}
\title{Create a background panel from a list of bam files (e.g. healthy samples)}
\usage{
create_background_panel(bam_list, targets, reference,
  vaf_threshold = 0.05, bam_list_tags = rep("", length(bam_list)),
  min_base_quality = 10, max_depth = 1e+05, min_mapq = 20,
  substitution_specific = T)
}
\arguments{
\item{bam_list}{A character vector of paths to bam files.}

\item{targets}{The targets data frame must have the columns chr, start and end.}

\item{reference}{The reference genome as BSgenome object.}

\item{vaf_threshold}{Loci with the fraction of non-reference reads above this value are masked with NA.}

\item{bam_list_tags}{RG tags for the list of bam files. By default, the whole bam file will be used.}

\item{min_base_quality}{The minimum base quality to count a read for a loci.}

\item{max_depth}{Maximum depth for the pileup}

\item{min_mapq}{The minimum mapping quality to count a read for a loci}

\item{substitution_specific}{logical, whether to have the loci by substitutions.}
}
\value{
A named list having depth, alt and vaf data frames. Each has the same order of loci in rows and the input samples in columns.
}
\description{
This function scans the targets regions in the list of bam files, and reports the number of reference, non-reference reads for each loci
in addition to the non-reference (VAF) allele frequency. Loci with VAF higher than vaf_threshold are masked with NA.
}
