% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_black_list.R
\name{create_black_list}
\alias{create_black_list}
\title{A function to create a black list of genomic loci based on a background panel created from a list of bam files (e.g. healthy samples)}
\usage{
create_black_list(background_panel, mean_vaf_quantile = 0.95,
  min_samples_one_read = max(2, ceiling(ncol(background_panel$vaf) *
  0.75)), min_samples_two_reads = max(2,
  ceiling(ncol(background_panel$vaf) * 0.2)))
}
\arguments{
\item{background_panel}{A list produced by create_background panel function}

\item{mean_vaf_quantile}{The quantile of mean VAF above which the loci are considered noisy}

\item{min_samples_one_read}{Loci that at least this number of samples exhibit at least one non-reference reads are considered noisy.}

\item{min_samples_two_reads}{Loci that at least this number of samples exhibit at least two non-reference reads are considered noisy.}
}
\value{
a character vector of the loci in the black list
}
\description{
The function applies criteria on the background panel to extract the noisy genomic loci. Criteria include minimum number of samples having
at least one and at least two non-reference allele. Additionally the quantile of mean VAF above which the loci are considered noisy
}
\seealso{
\code{\link{create_background_panel}}  \code{\link{test_ctDNA}}
}
