% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trinucleotide_context.R
\name{extract_trinucleotide_context}
\alias{extract_trinucleotide_context}
\title{A function to extract the trinucleotide context for a set of mutations}
\usage{
extract_trinucleotide_context(mutations, reference, destrand = T)
}
\arguments{
\item{mutations}{A data frame having the mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{reference}{the reference genome in BSgenome format}

\item{destrand}{logical, whether to destrand mutations}
}
\value{
A data frame with two columns having the substitutions and the trinucleotide context
}
\description{
A function to extract the trinucleotide context for a set of mutations
}
