% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mutations.R
\name{filter_mutations}
\alias{filter_mutations}
\title{Filter a set of mutations}
\usage{
filter_mutations(mutations, bams, tags = rep("", length(bams)),
  min_alt_reads = 2, min_samples = 2, min_base_quality = 20,
  max_depth = 1e+05, include_indels = F, min_mapq = 30)
}
\arguments{
\item{mutations}{A data frame with the reporter mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{bams}{a vector of paths to bam files}

\item{tags}{a vector of the RG tags if the bam has more than one sample}

\item{min_alt_reads}{the threshold of read counts showing alternative allele for a sample to be counted}

\item{min_samples}{the threshold of number of samples above which the mutations is filtered}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{max_depth}{maximum depth above which sampling will happen}

\item{include_indels}{whether to include indels in the pileup}

\item{min_mapq}{the minimum mapping quality for a read to be counted}
}
\value{
a named list contains: ref, vector of read counts of the reference alleles, and
        alt, vector of read counts of the alternative allele
}
\description{
Filter a set of mutations given the prevalence of their mismatches in a set of bam files. Mutations that 
have more than min_alt_reads in more than min_samples will be removed.
}
