% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_background_rate.R
\name{get_background_rate}
\alias{get_background_rate}
\title{Compute the background mismatch rate from a bam file}
\usage{
get_background_rate(bam, targets, reference, vaf_threshold = 0.1,
  tag = "", min_base_quality = 20, max_depth = 1e+05,
  min_mapq = 30)
}
\arguments{
\item{bam}{path to bam file}

\item{targets}{a data frame with the target regions. Must have three columns: chr, start and end}

\item{reference}{the reference genome in BSgenome format}

\item{vaf_threshold}{the bases with higher than this VAF threshold will be ignored in the calculation (real mutations)}

\item{tag}{the RG tag if the bam has more than one sample}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{max_depth}{maximum depth above which sampling will happen}

\item{min_mapq}{the minimum mapping quality for a read to be counted}
}
\value{
a list containing the general mismatch rate and substitution-specific rates
}
\description{
Runs through the target regions base by base counting the mismatches. Then it divides sum(msimatches)/sum(depths) for all bases in the targets
}
