% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_counts.R
\name{get_bam_chr}
\alias{get_bam_chr}
\title{Helper function to extract chromosome names from a bam file}
\usage{
get_bam_chr(bam)
}
\description{
Helper function to extract chromosome names from a bam file
}
