% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bam_SM.R
\name{get_bam_SM}
\alias{get_bam_SM}
\title{Helper function to extract SM from a bam file}
\usage{
get_bam_SM(bam, tag = "")
}
\arguments{
\item{bam}{Path to the bam file}

\item{tag}{The RG tag for the bam file}
}
\description{
Helper function to extract SM from a bam file
}
\keyword{internal}
