% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_size.R
\name{get_fragment_size}
\alias{get_fragment_size}
\title{Get fragment lengths from bam}
\usage{
get_fragment_size(bam, mutations = NULL, tag = "", isProperPair = NA,
  low_bound = 1, high_bound = 400)
}
\arguments{
\item{bam}{path to bam file}

\item{mutations}{An optional data frame with mutations. Must have the columns CHROM, POS, REF, ALT.}

\item{tag}{the RG tag if the bam has more than one samplee}

\item{isProperPair}{a logical wheter to return only proper pairs (T), only improper pairs (F), or it does not matter (NA)}

\item{low_bound}{Integer with the lowest fragment length}

\item{high_bound}{Integer with the highest fragment length}
}
\value{
A data frame with the columns Sample (SM tag in bam), ID (read ID), size (fragment size), and category (only if mutations is provided)
}
\description{
A function to extract fragment lengths from a bam file. Optionally, given a mutation data frame, it can categorize read lengths 
in mutated vs non-mutated reads.
}
