% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_fragment_size.R
\name{get_hist_bins}
\alias{get_hist_bins}
\title{Helper function to bin a variable}
\usage{
get_hist_bins(x, from, to, by, normalized)
}
\arguments{
\item{x}{the variable to be binned}

\item{from}{the minimum range}

\item{to}{the maximum range}

\item{by}{the break length}
}
\value{
A numeric vector having counts within bins normalized by the sum of the variable
}
\description{
Helper function to bin a variable
}
