% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_read_names.R
\name{get_mutation_read_names}
\alias{get_mutation_read_names}
\title{Get names of the reads showing alternative allele of a mutation}
\usage{
get_mutation_read_names(bam, chr, pos, alt, tag = "")
}
\arguments{
\item{bam}{path to bam file}

\item{chr}{Chromosome name for the mutation}

\item{pos}{Chromosome position for the mutation}

\item{alt}{The alternative allele of the mutation}

\item{tag}{the RG tag if the bam has more than one samplee}
}
\value{
A character vector having the read names
}
\description{
Extract the names of the reads in a bam file that support the variant allele of a single mutation
}
