% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutations_read_names.R
\name{get_mutations_read_names}
\alias{get_mutations_read_names}
\title{Get names of the reads showing alternative allele of a mutation}
\usage{
get_mutations_read_names(bam, mutations, min_base_quality = 20,
  tag = "", min_mapq = 30)
}
\arguments{
\item{bam}{path to bam file}

\item{mutations}{A data frame containing the mutations. Must have the columns CHROM, POS, REF, ALT.}

\item{min_base_quality}{integer specifying the minimum base quality for reads to be included.}

\item{tag}{the RG tag if the bam has more than one samplee}

\item{min_mapq}{integer specifying the minimum mapping quality for reads to be included}

\item{what}{Either Alt: names of the reads that exhibit variant alleles in mutations, 
or Ref: names of the reads the exhibit reference alleles in mutations}
}
\value{
A list with length equal to the number of mutations. Each element is a character vector with the read names.
}
\description{
Extract the names of the reads in a bam file that support the variant allele of a single mutation
}
