% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadCounts.R
\name{getReadCounts}
\alias{getReadCounts}
\title{Get read counts for a specific base in the genome}
\usage{
getReadCounts(chr, pos, base, bam, tag = "", min_base_quality = 20,
  max_depth = 1e+05, include_indels = F, min_mapq = 30)
}
\arguments{
\item{chr}{chromosome name}

\item{pos}{genomic coordinate}

\item{base}{the basepair for which count the reads}

\item{bam}{path to bam file}

\item{tag}{the RG tag if the bam has more than one sample}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{max_depth}{maximum depth above which sampling will happen}

\item{include_indels}{whether to include indels in the pileup}

\item{min_mapq}{the minimum mapping quality for a read to be counted}
}
\value{
a scalar, number of reads matching the base in the specified genomic position
}
\description{
Uses samtools pileup to get the read counts of the genomic position and the base specified
}
