% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mutations_in_phase.R
\name{merge_mutations_in_phase}
\alias{merge_mutations_in_phase}
\title{Collapses mutations in phase into one event}
\usage{
merge_mutations_in_phase(mutations, bam, tag = "",
  ID_column = "phasingID", min_base_quality = 20,
  use_unique_molecules = T)
}
\arguments{
\item{mutations}{A data frame with the reporter mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{bam}{path to bam file}

\item{tag}{the RG tag if the bam has more than one sample}

\item{ID_column}{The name of the column in mutations data.frame that has the IDs for mutations in phase.
NA values will be filled automatically by unique mutation identifiers.}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{use_unique_molecules}{A logical. If true, reads mapping to multiple mutations will 
be counted only once for the mutation that appears first in mutations data frame.}
}
\value{
A data frame that has the ref and alt counts for the mutations/events, as well as, 
the count of reads that support multiple mutations in phase, and the total number of reads.
}
\description{
Given a mutations data frame and a bam file, this function collapses mutations in phase identified by the ID_column into one event.
While doing that, it ignores the reads that support both the reference and alternative alleles for different mutations in phase.
When use_unique_molecules is TRUE, it counts the unique molecules only once, i.e. if two mutations are close to each others, the
counts for the second mutation include only the reads that do not map to the first one.
}
