% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_size.R
\name{plot_density}
\alias{plot_density}
\title{Plot density of variable}
\usage{
plot_density(df, var = "size", color = "Sample", binwidth = 2.5)
}
\arguments{
\item{df}{data frame in long format}

\item{var}{the column of which to plot the histogram}

\item{color}{the column containing the categories}

\item{binwidth}{bin width of the histogram}
}
\value{
a ggplot object
}
\description{
Plot density of variable
}
