% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivityTest.R
\name{positivityTest}
\alias{positivityTest}
\title{Compute a p-value of ctDNA positivity}
\usage{
positivityTest(depths, altReads, rate, seed, nPermutation)
}
\arguments{
\item{depths}{a vector with the depths of the variants}

\item{altReads}{the observed variant allele reads}

\item{rate}{the mismatch rate}

\item{seed}{the random seed}

\item{nPermutation}{the number of permutations to run.}
}
\value{
a scalar, permutation p-value.
}
\description{
A function to determine ctDNA positivity with N repeated permutations. Return a p-value based on permutation test. Calls simulator function  nPermutation times.
}
