% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_fragment_size.R
\name{sample_proportion}
\alias{sample_proportion}
\title{Helper function to sample proportion of a variable}
\usage{
sample_proportion(x, proportion = 0.5, seed = 42)
}
\arguments{
\item{x}{the variable to sample from.}

\item{proportion}{the fraction of the variable to be randomly sample.}

\item{seed}{the random seed.}
}
\value{
a vector having randomly sampled proportion of the input variable.
}
\description{
Helper function to sample proportion of a variable
}
