% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{A function to sample from binomial distribution}
\usage{
simulator(nVariants, depths, rate, altReads, seed)
}
\arguments{
\item{nVariants}{the number of variants to simulate.}

\item{depths}{a vector with the depths of the variants}

\item{rate}{the mismatch rate}

\item{altReads}{the observed variant allele reads}

\item{seed}{the random seed}
}
\value{
a scalar. Either 1 if the simulation exceedds observed variant alleles or 0 otherwise
}
\description{
Samples from binomial distribution N variants with different depths and fixed mismatch rate for one permutation. Return 1 if the simulation exceeds observed alt allele reads or 0 otherwise
}
