% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fragment_size.R
\name{summarize_fragment_size}
\alias{summarize_fragment_size}
\title{The function summarizes fragment size in defined genomic regions}
\usage{
summarize_fragment_size(bam, regions, tag = "",
  summary_functions = list(Mean = mean, Median = median), ...)
}
\arguments{
\item{bam}{the input bam file}

\item{regions}{data frame contatining the genomic regions. Must have the columns chr, start and end.}

\item{tag}{the RG tag if the bam has more than one samplee.}

\item{summary_functions}{a named list containing the R functions used for summarization, e.g. mean, sd.}

\item{...}{Other parameters passed to get_fragment_size}
}
\value{
a data frame with the first column having the regions in the format of chr:start-end, and other columns correspond to summary_functions.
}
\description{
The function summarizes fragment size in defined genomic regions
}
