% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targets}
\alias{targets}
\title{Example Genomic targets to use with ctDNAtools package}
\format{A data frame with 1 row and 3 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start genomic position}
  \item{end}{end genomic position}
}}
\source{
In-house
}
\usage{
targets
}
\description{
Includes 1 target region in chr14.
}
\keyword{datasets}
