% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ctDNA.R
\name{test_ctDNA}
\alias{test_ctDNA}
\title{Tests the ctDNA positivity of a sample}
\usage{
test_ctDNA(mutations, bam, targets, reference, tag = "",
  ID_column = NULL, black_list = NULL, substitution_specific = T,
  vaf_threshold = 0.1, min_base_quality = 30, max_depth = 1e+05,
  min_mapq = 40, bam_list = NULL, bam_list_tags = rep("",
  length(bam_list)), min_alt_reads = 1,
  min_samples = ceiling(length(bam_list)/10), by_substitution = F,
  n_simulations = 10000, pvalue_threshold = 0.05, seed = 123,
  informative_reads_threshold = 10000)
}
\arguments{
\item{mutations}{A data frame with the reporter mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{bam}{path to bam file}

\item{targets}{a data frame with the target regions. Must have three columns: chr, start and end}

\item{reference}{the reference genome in BSgenome format}

\item{tag}{the RG tag if the bam has more than one sample}

\item{ID_column}{The name of the column that contains the ID of mutations in phase. All mutations in Phase should have the same ID in that column. 
Will lead to considerable slow down when provided.}

\item{black_list}{a character vector of genomic loci to filter. The format is chr_pos if substitution_specific is false or
chr_pos_ref_alt if substitution_specific is true. If given, will override bam_list.}

\item{substitution_specific}{logical, whether to have the loci of black_list by substitutions.}

\item{vaf_threshold}{the bases with higher than this VAF threshold will be ignored in the calculation (real mutations)}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{max_depth}{maximum depth above which sampling will happen}

\item{min_mapq}{the minimum mapping quality for a read to be counted}

\item{bam_list}{A vector containing the paths to bam files used to filter mutations. Mutations that have more than min_alt_reads in more than min_samples will be filtered.}

\item{bam_list_tags}{the RG tags for the bams included in bams list.}

\item{min_alt_reads}{When bam_list is provided, this sets the minimum number of alternative allele reads for a sample to be counted.}

\item{min_samples}{When number of samples having more than min_alt_reads exceeds this number, the mutation will be filtered.}

\item{by_substitution}{boolean whether to run the test according to substitution-specific background rate}

\item{n_simulations}{the number of simulations.}

\item{pvalue_threshold}{the p-value threshold used to decide positivity or negativity.}

\item{seed}{the random seed.}

\item{informative_reads_threshold}{the number of informative reads (unique reads mapping to specified mutations) under which the test will be undetermined.}
}
\value{
a named list contains: counts, a data frame of read counts of reference and variant alleles for the reporter mutations in the tested sample,
        backgroundRate, a list of substituion-specific background rate, and pvalue, the p-value of the test
}
\description{
Given a set of reporter mutation, this functions counts the reads matching the reporter mutations in the sample to be tested, 
estimates the mismatch rate for the sample to be tested, and then runs a Monte Carlo simulation test to determine whether the tested sample is positive or negative.
}
