% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_tag.R
\name{verify_tag}
\alias{verify_tag}
\title{Helper function to verify a tag exists}
\usage{
verify_tag(bam, tag)
}
\arguments{
\item{bam}{path to bam file}

\item{tag}{the RG tag if the bam has more than one sample}
}
\value{
logical, whether tag in bam
}
\description{
Helper function to verify a tag exists
}
\keyword{internal}
