% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_fragment_size.R
\name{bin_fragment_size}
\alias{bin_fragment_size}
\title{Get histogram density of fragment lengths from a bam file}
\usage{
bin_fragment_size(bam, mutations = NULL, tag = "", bin_size = 2,
  mutated_only = F, augmented = F, n_augmentation = 10,
  augment_proportion = 0.5, seed = 123, isProperPair = NA,
  min_size = 1, max_size = 400, ignore_trimmed = T,
  different_strands = T, simple_cigar = F)
}
\arguments{
\item{bam}{path to bam file.}

\item{mutations}{An optional data frame with mutations. Must have the columns CHROM, POS, REF, ALT.}

\item{tag}{the RG tag if the bam has more than one samplee.}

\item{bin_size}{the width of the bin (breaks) of the histogram.}

\item{mutated_only}{A logical, whether to return the counts for only mutated reads. The 'mutations' input should be given when TRUE.}

\item{augmented}{A logical, whether to create additional samples by sampling proportion of the reads.}

\item{n_augmentation}{Number of pseudo-samples to create.}

\item{augment_proportion}{the fraction of the reads to sample in each augmented sample.}

\item{seed}{The random seed to use when creating augmented samples.}

\item{isProperPair}{a logical wheter to return only proper pairs (T), only improper pairs (F), or it does not matter (NA).}

\item{min_size}{Integer with the lowest fragment length.}

\item{max_size}{Integer with the highest fragment length.}

\item{ignore_trimmed}{logical, whether to remove reads that have been hard trimmed.}

\item{different_strands}{logical, whether to keep only reads whose mates map to different strand.}

\item{simple_cigar}{logical, whether to include only reads with simple cigar.}
}
\value{
A data frame with one column corresponding to the sample name from the bam. 
Each row has the count of fragment lengths within the bin normalized by the total number of reads.
}
\description{
The function first extracts fragment length from a bam file then computes the histogram over defined bins. Optionally,
it can computes the histogram of fragment lengths only for mutated reads (confirmed ctDNA molecules).
}
