% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_size.R
\name{get_fragment_size}
\alias{get_fragment_size}
\title{Get fragment lengths from bam}
\usage{
get_fragment_size(bam, mutations = NULL, tag = "", isProperPair = NA,
  mapqFilter = 30, min_size = 1, max_size = 400,
  ignore_trimmed = T, different_strands = T, simple_cigar = F)
}
\arguments{
\item{bam}{path to bam file.}

\item{mutations}{An optional data frame with mutations. Must have the columns CHROM, POS, REF, ALT.}

\item{tag}{the RG tag if the bam has more than one samplee.}

\item{isProperPair}{a logical wheter to return only proper pairs (T), only improper pairs (F), or it does not matter (NA).}

\item{mapqFilter}{mapping quality threshold for considering reads.}

\item{min_size}{Integer with the lowest fragment length.}

\item{max_size}{Integer with the highest fragment length.}

\item{ignore_trimmed}{logical, whether to remove reads that have been hard trimmed.}

\item{different_strands}{logical, whether to keep only reads whose mates map to different strand.}

\item{simple_cigar}{logical, whether to include only reads with simple cigar.}
}
\value{
A data frame with the columns Sample (SM tag in bam), ID (read ID), size (fragment size), and category (only if mutations is provided).
}
\description{
A function to extract fragment lengths from a bam file. Optionally, given a mutation data frame, it can categorize read lengths 
in mutated vs non-mutated reads.
}
