% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hist_bins.R
\name{get_hist_bins}
\alias{get_hist_bins}
\title{Helper function to bin a variable}
\usage{
get_hist_bins(x, from, to, by, normalized, custom_bins = NULL)
}
\arguments{
\item{x}{the variable to be binned}

\item{from}{the minimum range}

\item{to}{the maximum range}

\item{by}{the break length}

\item{normalized}{whether to normalize the counts to total counts.}

\item{custom_bins}{A numeric vector for custom breaks to bin the histogram of fragment length. Over-rides bin_size.}
}
\value{
A numeric vector having counts within bins normalized by the sum of the variable
}
\description{
Helper function to bin a variable
}
\keyword{internal}
