% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mutations_in_phase.R
\name{merge_mutations_in_phase}
\alias{merge_mutations_in_phase}
\title{Collapses mutations in phase into one event}
\usage{
merge_mutations_in_phase(mutations, bam, tag = "",
  ID_column = "phasingID", min_base_quality = 20, min_mapq = 30)
}
\arguments{
\item{mutations}{A data frame with the reporter mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{bam}{path to bam file}

\item{tag}{the RG tag if the bam has more than one sample}

\item{ID_column}{The name of the column in mutations data.frame that has the IDs for mutations in phase.
NA values will be filled automatically by unique mutation identifiers.}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{min_mapq}{integer specifying the minimum mapping quality for reads to be included.}
}
\value{
A data frame that has the ref and alt counts for the mutations/events, as well as, 
the count of reads that support multiple mutations in phase, and the total number of reads.
}
\description{
Given a mutations data frame and a bam file, this function collapses mutations in phase identified by the ID_column into one event.
While doing that, it ignores the reads that support both the reference and alternative alleles for different mutations in phase.
}
