% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity_test.R
\name{positivity_test}
\alias{positivity_test}
\title{Compute a p-value of ctDNA positivity}
\usage{
positivity_test(depths, altReads, rate, seed, n_permutation = 10000)
}
\arguments{
\item{depths}{a vector with the depths of the variants}

\item{altReads}{the observed variant allele reads}

\item{rate}{the mismatch rate}

\item{seed}{the random seed}

\item{n_permutation}{the number of permutations to run.}
}
\value{
a scalar, permutation p-value.
}
\description{
A function to determine ctDNA positivity with N repeated permutations. Return a p-value based on permutation test. Calls simulator function  n_permutation times.
}
