% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fragment_size.R
\name{summarize_fragment_size}
\alias{summarize_fragment_size}
\title{The function summarizes fragment size in defined genomic regions}
\usage{
summarize_fragment_size(bam, regions, tag = "", isProperPair = NA,
  mapqFilter = 30, different_strands = T, min_size = 1,
  max_size = 400, simple_cigar = F, ignore_trimmed = T,
  summary_functions = list(Mean = mean, Median = median))
}
\arguments{
\item{bam}{the input bam file}

\item{regions}{data frame contatining the genomic regions. Must have the columns chr, start and end.}

\item{tag}{the RG tag if the bam has more than one samplee.}

\item{isProperPair}{a logical wheter to return only proper pairs (T), only improper pairs (F), or it does not matter (NA).}

\item{mapqFilter}{mapping quality threshold for considering reads.}

\item{different_strands}{logical, whether to keep only reads whose mates map to different strand.}

\item{min_size}{Integer with the lowest fragment length.}

\item{max_size}{Integer with the highest fragment length.}

\item{simple_cigar}{logical, whether to include only reads with simple cigar.}

\item{ignore_trimmed}{logical, whether to remove reads that have been hard trimmed.}

\item{summary_functions}{a named list containing the R functions used for summarization, e.g. mean, sd.}
}
\value{
a data frame with the first column having the regions in the format of chr:start-end, and other columns correspond to summary_functions.
}
\description{
The function summarizes fragment size in defined genomic regions
}
