% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ctDNA.R
\name{test_ctDNA}
\alias{test_ctDNA}
\title{Tests the ctDNA positivity of a sample}
\usage{
test_ctDNA(mutations, bam, targets, reference, tag = "",
  vafThreshold = 0.1, min_base_quality = 20, max_depth = 10000,
  min_mapq = 30, nPermutation = 10000, seed = 123)
}
\arguments{
\item{mutations}{A data frame with the reporter mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{bam}{path to bam file}

\item{targets}{a data frame with the target regions. Must have three columns: chr, start and end}

\item{reference}{the reference genome in BSgenome format}

\item{tag}{the RG tag if the bam has more than one sample}

\item{vafThreshold}{the bases with higher than this VAF threshold will be ignored in the calculation (real mutations)}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{max_depth}{maximum depth above which sampling will happen}

\item{min_mapq}{the minimum mapping quality for a read to be counted}

\item{nPermutation}{the number of permutations}

\item{seed}{the random seed}
}
\value{
a named list contains: counts, a data frame of read counts of reference and variant alleles for the reporter mutations in the tested sample,
        backgroundRate, a scalar estimating background rate, and pvalue, the p-value of the test
}
\description{
Given a set of reporter mutation, this functions counts the reads matching the reporter mutations in the sample to be tested, 
estimates the mismatch rate for the sample to be tested, and then runs a permutation test to determine whether the tested sample is positive or negative.
}
