% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ctDNA.R
\name{test_ctDNA}
\alias{test_ctDNA}
\title{Tests the ctDNA positivity of a sample}
\usage{
test_ctDNA(mutations, bam, targets, reference, tag = "",
  ID_column = NULL, use_unique_molecules = T, vaf_threshold = 0.1,
  min_base_quality = 20, max_depth = 1e+05, min_mapq = 30,
  bam_list = character(), bam_list_tags = rep("", length(bam_list)),
  min_alt_reads = 1, min_samples = 1, by_substitution = F,
  n_simulations = 10000, seed = 123)
}
\arguments{
\item{mutations}{A data frame with the reporter mutations. Should have the columns CHROM, POS, REF, ALT.}

\item{bam}{path to bam file}

\item{targets}{a data frame with the target regions. Must have three columns: chr, start and end}

\item{reference}{the reference genome in BSgenome format}

\item{tag}{the RG tag if the bam has more than one sample}

\item{ID_column}{The name of the column that contains the ID of mutations in phase. All mutations in Phase should have the same ID in that column. 
Will lead to considerable slow down when provided.}

\item{use_unique_molecules}{A logical. If true, reads mapping to multiple mutations will 
be counted only once for the mutation that appears first in mutations data frame.}

\item{vaf_threshold}{the bases with higher than this VAF threshold will be ignored in the calculation (real mutations)}

\item{min_base_quality}{minimum base quality for a read to be counted}

\item{max_depth}{maximum depth above which sampling will happen}

\item{min_mapq}{the minimum mapping quality for a read to be counted}

\item{bam_list}{A vector containing the paths to bam files used to filter mutations. Mutations that have more than min_alt_reads in more than min_samples will be filtered.}

\item{bam_list_tags}{the RG tags for the bams included in bams list.}

\item{min_alt_reads}{When bam_list is provided, this sets the minimum number of alternative allele reads for a sample to be counted.}

\item{min_samples}{When number of samples having more than min_alt_reads exceeds this number, the mutation will be filtered.}

\item{by_substitution}{boolean whether to run the test according to substitution-specific background rate}

\item{n_simulations}{the number of simulations}

\item{seed}{the random seed}
}
\value{
a named list contains: counts, a data frame of read counts of reference and variant alleles for the reporter mutations in the tested sample,
        backgroundRate, a list of substituion-specific background rate, and pvalue, the p-value of the test
}
\description{
Given a set of reporter mutation, this functions counts the reads matching the reporter mutations in the sample to be tested, 
estimates the mismatch rate for the sample to be tested, and then runs a Monte Carlo simulation test to determine whether the tested sample is positive or negative.
}
